requirements.txt
asn1crypto==1.4.0
attrs==21.4.0
boto3==1.21.3
botocore==1.24.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.1.0
moto==3.0.4
numpy==1.22.2
oscrypto==1.2.1
packaging==21.3
pandas==1.3.5
pluggy==1.0.0
py==1.11.0
pyarrow==6.0.1
pycparser==2.21
pycryptodomex==3.14.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
responses==0.18.0
s3transfer==0.5.1
six==1.16.0
snowflake-connector-python==2.7.4
tomli==2.0.1
urllib3==1.26.8
Werkzeug==2.0.3
xmltodict==0.12.0
